WIP: Earning experience
Done:
- Only show floating cube and menu when villager has BOP
- Confirm BOP gets unlocked at level boundary

TODO:
- Villager should take BOP to town flag
- BOP "inventory" UI on town flag
- Test all commands
