+++ src/main/java/ca/bradj/questown/town/interfaces/VillagerHolder.java
@@ -104,4 +104,6 @@ public interface VillagerHolder {
             UUID uuid,
             int exp
     );
+
+    boolean hasBlockOfProgress(UUID uuid);
 }
